+++ src/main/java/com/pehulja/threading/counters/plain/PlainCounter.java
@@ -3,7 +3,7 @@ package com.pehulja.threading.counters.plain;
 public class PlainCounter extends AbstractPlainCounter{
 
 	@Override
-	public int incrementAndGet() {
+	public long incrementAndGet() {
 		return ++counter;
 	}
 }
