+++ src/main/java/com/pehulja/threading/counters/plain/SynchronizedMethodPlainCounter.java
@@ -2,7 +2,7 @@ package com.pehulja.threading.counters.plain;
 
 public class SynchronizedMethodPlainCounter extends AbstractPlainCounter{
 	@Override
-	public synchronized int incrementAndGet() {
+	public synchronized long incrementAndGet() {
 		return ++counter;
 	}
 }
